myproject/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2.1
django-ninja==1.4.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.13.2
